move function to custom report utils
